+++ pkg/skaffold/trigger/triggers.go
@@ -27,7 +27,7 @@ import (
 	"time"
 
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/color"
-	runcontext "github.com/GoogleContainerTools/skaffold/pkg/skaffold/runner/context"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/runner/runcontext"
 	"github.com/rjeczalik/notify"
 	"github.com/sirupsen/logrus"
 )
